requirements.txt
dgl-cu111==0.6.1   # For CUDA 
# dgl==0.6.1       # For CPU
biopython==1.79
matplotlib==3.5.2
numpy==1.21.5
pandas==1.3.5
scikit-learn==0.24.2
scipy==1.7.3
torch==1.10.0
transformers==4.20.1
zipfile37==0.1.3
bio-embeddings[transformers]==0.2.2